+++ matrix/Matrix_Pointers.hxx
@@ -63,15 +63,15 @@ namespace Seldon
     void Clear();
 
     // Basic methods.
-    int GetDataSize() const;
+    size_t GetDataSize() const;
     int64_t GetMemorySize() const;
     pointer* GetMe() const;
 
     // Memory management.
-    void Reallocate(int i, int j);
+    void Reallocate(size_t i, size_t j);
     void SetData(int i, int j, pointer data);
     void Nullify();
-    void Resize(int i, int j);
+    void Resize(size_t i, size_t j);
 
     // Element access and affectation.
     pointer GetDataPointer(int i, int j) const;
